Add (wip) Sponsors section
Replacing 'registration' with sponsors. Added brief text into the
location section explaining how to view the conference and info that
registration is not a thing for this year.

Added github ribbon and footnote.

Fix missing nav quirk when view width is exactly 1080 pixels. Thank you
to samueldr for pointing that out.